rename folder home and cart, create file form scss
