+++ src/stylesheets/modules/_all.scss
@@ -3,3 +3,4 @@
 @import 'button';
 @import 'product';
 @import 'cart';
+@import 'form';
