+++ src/stylesheets/modules/_form.scss
@@ -1,0 +1,28 @@
+.form {
+  max-width: 670px;
+  margin: 175px auto;
+  border-radius: 8px;
+  background: map-get($color, 'white');
+  box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
+  font-size: map-get($font-size, 'md-2-5');
+  color: map-get($color, text);
+  &-detail {
+    padding: 30px 45px;
+  }
+  &-title {
+    position: relative;
+    padding: 3px 0 20px;
+    margin-bottom: 40px;
+    font-weight: map-get($font-weight, 'bold');
+    font-size: map-get($font-size, 'lg');
+    &::after {
+      content: "";
+      position: absolute;
+      top: 100%;
+      left: 50%;
+      @include size(50px, 2px);
+      background-color: map-get($color, primary);
+      transform: translateX(-50%);
+    }
+  }
+}
